# openapi/homechef-api.yaml
openapi: 3.0.3
info:
  title: HomeChef API
  description: |
    Comprehensive API for HomeChef - A platform connecting home chefs with customers for authentic homemade meals.
    
    ## Features
    - Multi-role authentication (Customer, Chef, Delivery Partner, Admin)
    - Complete order management lifecycle
    - Real-time order tracking
    - Payment processing with multiple methods
    - Review and rating system
    - Analytics and reporting
    - Push notifications
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Authenticated users: 1000 requests/hour
    - Unauthenticated: 100 requests/hour
    - Admin users: 5000 requests/hour
    
  version: 1.0.0
  contact:
    name: HomeChef API Support
    email: api-support@homechef.com
    url: https://docs.homechef.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.homechef.com/v1
    description: Production server
  - url: https://staging-api.homechef.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                role:
                  type: string
                  enum: [customer, chef, delivery, admin]
                  example: customer
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/send-otp:
    post:
      tags: [Authentication]
      summary: Send OTP
      description: Send OTP to phone number for verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, role]
              properties:
                phone:
                  type: string
                  pattern: '^\+91[6-9]\d{9}$'
                  example: '+919876543210'
                role:
                  type: string
                  enum: [customer, chef, delivery]
                  example: customer
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPResponse'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/verify-otp:
    post:
      tags: [Authentication]
      summary: Verify OTP
      description: Verify OTP and complete authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp, role]
              properties:
                phone:
                  type: string
                  example: '+919876543210'
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
                role:
                  type: string
                  enum: [customer, chef, delivery]
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # User Management
...
  /chefs/{chef_id}/menu:
    get:
      tags: [Menu Management]
      summary: Get chef menu
      description: Retrieve menu items for a specific chef
      security: []
      parameters:
        - $ref: '#/components/parameters/ChefId'
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            example: 'main_course'
        - name: available
          in: query
          description: Filter by availability
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Menu retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'

  # Order Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ChefId:
      name: chef_id
      in: path
      required: true
      description: Chef ID
      schema:
        type: string
        example: 'chef_123'
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation successful'
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: object
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
                refresh_token:
                  type: string
                  example: 'refresh_token_here'
                expires_in:
                  type: integer
                  example: 3600
    OTPResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                otp_id:
                  type: string
                  example: 'otp_123'
                expires_in:
                  type: integer
                  example: 300
    User:
      type: object
      properties:
        id:
          type: string
          example: 'user_123'
        email:
          type: string
          format: email
          example: 'user@example.com'
        name:
          type: string
          example: 'John Doe'
        phone:
          type: string
          example: '+919876543210'
        role:
          type: string
          enum: [customer, chef, delivery, admin]
          example: 'customer'
        status:
          type: string
          enum: [active, inactive, suspended]
          example: 'active'
        email_verified:
          type: boolean
          example: true
        phone_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
...
    MenuItem:
      type: object
      properties:
        id:
          type: string
          example: 'dish_123'
        name:
          type: string
          example: 'Butter Chicken'
        description:
          type: string
          example: 'Rich and creamy tomato-based curry with tender chicken pieces'
        price:
          type: number
          format: float
          example: 280.00
        category:
          type: string
          example: 'main_course'
        cuisine_type:
          type: string
          example: 'north_indian'
        images:
          type: array
          items:
            type: string
            format: uri
          example: ['https://cdn.homechef.com/dishes/dish_123_1.jpg']
        is_vegetarian:
          type: boolean
          example: false
        is_vegan:
          type: boolean
          example: false
        spice_level:
          type: string
          enum: [mild, medium, hot]
          example: 'medium'
        preparation_time:
          type: integer
          description: 'Preparation time in minutes'
          example: 25
        serves:
          type: integer
          example: 2
        ingredients:
          type: array
          items:
            type: string
          example: ['chicken', 'tomatoes', 'cream', 'spices']
        allergens:
          type: array
          items:
            type: string
          example: ['dairy']
        nutritional_info:
          type: object
          properties:
            calories:
              type: integer
              example: 420
            protein:
              type: number
              format: float
              example: 28.0
            carbs:
              type: number
              format: float
              example: 12.0
            fat:
              type: number
              format: float
              example: 32.0
        is_available:
          type: boolean
          example: true
        rating:
          type: number
          format: float
          example: 4.8
        total_orders:
          type: integer
          example: 156
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'BAD_REQUEST'
              message: 'Invalid request format'
            timestamp: '2024-01-15T10:30:00Z'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Authentication required'
            timestamp: '2024-01-15T10:30:00Z'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'VALIDATION_ERROR'
              message: 'Validation failed'
              details:
                email: 'Invalid email format'
                password: 'Password must be at least 6 characters'
            timestamp: '2024-01-15T10:30:00Z'
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'RATE_LIMIT_EXCEEDED'
              message: 'Too many requests'
            timestamp: '2024-01-15T10:30:00Z'
...